coerce factors to correct data type

+++ train/train.py
@@ -37,7 +37,7 @@ def train_epoch_verbose(model,optimizer,loader,base_sequence,loss_function,rando
         data = data.to(model.device)
         optimizer.zero_grad()
         if conditional:
-            c = batch[1].to(model.device).view(1,-1)
+            c = batch[1].to(torch.float32).to(model.device).view(1,-1)
             samples = model(base_sequence,random,mod,c)
         else:
             samples = model(base_sequence,random,mod)
